requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.7.0
gradio==5.37.0
gradio_client==1.10.4
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
openai==1.96.1
opencv-python==4.12.0.88
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.3
safehttpx==0.1.6
safetensors==0.5.3
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
sympy==1.14.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.167
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
